requirements.txt
aiohttp>=3.3.2
attrs>=18.1.0
